:recycle: Refactor: Util (JSON & CSV)
