+++ zerohertzLib/util/__init__.py
@@ -1,5 +1,5 @@
+from zerohertzLib.util.csv import write_csv
 from zerohertzLib.util.data import MakeData
-from zerohertzLib.util.json import Json, JsonDir
-from zerohertzLib.util.write import write_csv, write_json
+from zerohertzLib.util.json import Json, JsonDir, write_json
 
 __all__ = ["Json", "JsonDir", "MakeData", "write_csv", "write_json"]
