+++ zerohertzLib/util/csv.py
@@ -30,40 +30,3 @@ def write_csv(data: List[List[Any]], path: str, tsv: Optional[bool] = False) -> 
             for d in data:
                 f.writelines(",".join(list(map(str, d))) + "\n")
         return os.path.abspath(f"{path}.csv")
-
-
-def write_json(data: Union[Dict[Any, Any], List[Dict[Any, Any]]], path: str) -> str:
-    """JSON (JavaScript Object Notation)를 작성하는 함수
-
-    Args:
-        data (``Dict[Any, Any]``): 입력 데이터 (header 포함 무관)
-        path (``str``): 출력될 json 파일의 경로 및 파일명
-
-    Returns:
-        ``str``: 파일의 절대 경로
-
-    Examples:
-        >>> zz.util.write_json([{"id": "4169", "전투력": 4209, "정보": ["아무", "거나"]}]*100, "zerohertzLib/star_craft")
-        '/.../star_craft.json'
-        [
-            {
-                "id": "4169",
-                "전투력": 4209,
-                "정보": [
-                    "아무",
-                    "거나"
-                ]
-            },
-            {
-                "id": "4169",
-                "전투력": 4209,
-                "정보": [
-                    "아무",
-                    "거나"
-                ]
-            },
-        ...
-    """
-    with open(f"{path}.json", "w", encoding="utf-8") as f:
-        json.dump(data, f, indent=4, ensure_ascii=False)
-    return os.path.abspath(f"{path}.json")
